refactor: solidify observer side of elections
Previously observation was done "magically" by listening to an event.
This made it hard to monitor status and lifecycle. This moves the logic
into a subclass. Additionally, revisions were not being maintained
between different queries introducing the possibility for race
conditions. The inner event loop is functionally identical to the
one in the etcd3 official client.